requirements.txt
# Django
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5
drf-spectacular>=0.27.0
drf-nested-routers

# WebSockets
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
websockets>=12.0,<14.0  # Pin to avoid transfer_data_task AttributeError

# Database
psycopg2-binary>=2.9.9

# Celery
celery>=5.3.4
redis>=5.0.1
django-celery-beat>=2.5.0

# HTTP Client
requests>=2.31.0

# Security
cryptography>=41.0.7

# Production
gunicorn>=21.2.0
uvicorn[standard]>=0.27.0
whitenoise>=6.6.0

# Development
python-dotenv>=1.0.0

# Export
openpyxl>=3.1.2
xlsxwriter>=3.1.9

# Payments
mercadopago>=2.3.0

# Email
resend>=0.7.0

# Storage (AWS S3)
django-storages>=1.14.0
boto3>=1.34.0

# Image processing
Pillow>=10.0.0

# Langchain & AI (Native - sem Langflow)
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-anthropic>=0.1.0
